agri_backend/requirements.txt
absl-py==2.2.0
annotated-types==0.7.0

arrow==1.3.0
asgiref==3.8.1

astunparse==1.6.3
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.2
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
filelock==3.18.0
Flask==2.3.1
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
h5py==3.13.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
joblib==1.4.2
jsonpointer==3.0.0
jupyter==1.1.1
jupyter-console==6.6.3
jupyterlab_widgets==3.0.13
kaggle==1.7.4.2
kagglehub==0.3.10
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.10.1
mdurl==0.1.2

ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8

networkx==3.4.2
nodeenv==1.9.1

numba==0.61.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.14.1

pandas==2.2.3
pillow==11.1.0
prisma==0.15.0

protobuf==5.29.4
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1

python-dotenv==1.0.1

python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.1
pywin32==308
rich==13.9.4

scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2

setuptools==75.8.0
shap==0.47.1

slicer==0.0.8

SQLAlchemy==2.0.39
sqlparse==0.5.1

starlette==0.46.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0

text-unidecode==1.3
threadpoolctl==3.6.0

tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0


types-python-dateutil==2.9.0.20241206

tzdata==2024.2
uri-template==1.3.0

uvicorn==0.34.0

webcolors==24.11.1
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
WTForms==3.2.1
